subscriber-profile-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
pyjwt[crypto]==2.9.0
prometheus-client==0.20.0
python-dotenv==1.0.1
python-multipart==0.0.9
